Simulate constant delta time and clear individual graphic elements
It is better to simulate each game update independent of the fps. For
this we are using a time accumulator to help us keep track of the time
between frames.

Profiling revealed a lot of time taken by the background screen redraw.
This optimization consists of clearing the space drawn by each element
before being drawn.